refactor: move authentication middleware to dedicated file
